Fix logcollector index in staging/prod
Summary:
I put out a previous diff that should fix the index comparison issue in log-collector, and tested on plc-dev.
after landing, i tried it out on staging and was surprised to see the log-collector was still complaining about reindexing.
looking at the elastic settings, i saw the log indices on plc-staging and plc both had 1 shard and 1 replica only, which was why the index comparison was still failing.
renaming the index once again so that we don't lose our previous logs, but we can start up an index with the correct settings.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6135

GitOrigin-RevId: 50bc0f28ba2e69ea451d6c8105eeaa53514fcf40