Fix bug in UI where Execution Stats tab didn't work when clicked
Summary: Fixing this to help with investigating streaming queries. When the active tab name wasn't set to a table name, we would reset it to the first table. this behavior didn't account for the fact that the stats tab was also a valid selection, so that logic is now incorporated.

Test Plan: ran the ui

Reviewers: michelle, philkuz, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6141

GitOrigin-RevId: 9d10c701a7a47e3bb074f3a192098e4165de957e