Run etcd Gets with serializable option
Summary: etcd reads are faster if we don't need to ensure linearizability. since we have a single replica, we shouldnt need linearizability

Test Plan:
ran skaffold and a bunch of queries... it doesnt seem to break anything.
testing the timing is harder since we should probably run a real benchmark for that

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6157

GitOrigin-RevId: ea207be6003ba62bba17de71d4cee10001e22e4c