PP-2115: Change TransferResultSink API to support result sinks initializing their connection to the downstream destination.
Summary:
Streaming queries may spend a long time between sending result batches if the data they are producing is sparse. That means that the timeout based approach of monitoring query health will not scale to streaming queries, because a timeout is no longer an accurate way of assessing if a query is healthy. We want to replace this timeout method with checking the health of GRPC connections from the sources to the destnations for the result data. Kelvin nodes will cancel their queries if the connection between them and their source data agents gets lost. The query broker already does this handling, if one of its agent streams gets disconnected, it will cancel the rest of the query everywhere else. However, the question remained for how a query would be cancelled if its connections from source to destination were never established in the first place. The solution here is to have all sinks send a message with their identity and table/result they are producing when establishing a stream to a remote destination. Then, the destination can keep track of which streams it has received initialization for, and timeout if a sink takes too long to set up its connection to the destination.

This diff sets the groundwork for that by changing up the TransferResultSink message to support sending an initial message for a stream initializing the result sink. Next diff will have the query broker expecting and using those initialization messages, and removing its timeouts. After that, exec_graph.cc/h will be refactored to no longer timeout when data has not come in in a while, but do timeout if any of the exec graphs remote sources have taken too long to establish a connection.

Test Plan: existing/unit

Reviewers: michelle, philkuz, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2115

Differential Revision: https://phab.corp.pixielabs.ai/D6167

GitOrigin-RevId: 8a24e118e58bbdcf07f9b57efff29eeb75d8ed09