Rm updatedAgentsMutex dependency on etcd calls
Summary:
whenever we have a lock, it should not have to wait on particularly slow operations, such as etcd.
i updated most of the wrappers so that the lock is acquired after the metadata operation.

@nserrino, you probably know this area best:
the one that probably needs most review is the GetAgentUpdates fix, where we pull out the updates and clear the updatedAgents before making metadata calls.
one condition I could imagine that can happen now is:
- readInitialState is true
- we start to read the agents/agentData from etcd
- in the middle, we have other agent updates which are written to updatedAgents
- we may or may not read these new updates from etcd, depending on timing
- the next time GetAgentUpdates is called, we may read updates in updatedAgents that were already sent previously. will the query broker accept this?

Test Plan: ran in skaffold and unit tests, things seem to still work

Reviewers: nserrino, zasgar, #engineering

Reviewed By: nserrino, zasgar, #engineering

Subscribers: nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D6152

GitOrigin-RevId: 0f55a30d257156f98053c227868b99f2e32f91cf