Downgrade inconsistent # of agent exec stats from error to log
Summary: Thought that this should be an error condition, but based on the control flow it can be correct that an agent will fail to send exec stats before the kelvin completes the query. In addition, an agent may go away in the middle of the query and that is not an error.

Test Plan: n/a

Reviewers: zasgar, #engineering, philkuz

Reviewed By: #engineering, philkuz

Subscribers: philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D6159

GitOrigin-RevId: 64f54b196b1a16a3f98164f77cff077ed6a4e2bf