Update etcd snapshot count
Summary:
we're still running out of space on customer's etcd, except now metadata is reporting that the etcd data itself is only 600 mb...
we need to figure out what else is taking up space.

a higher snapshot count increases the amount of snapshots we take, but reduces the number of raft logs we need to hold in memory.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6170

GitOrigin-RevId: e29cd8642e6c8911e555e13aa6dd5c9ffe8df6ea