Fix LaunchDarkly identify bug
Summary:
we were seeing "TypeError: Cannot read property 'identify' of undefined"
this is because the withLDProvider that sets up the client is running synchronously and can sometimes take longer to load than the vizier page renders.
to fix this we could use asyncWithLDProvider, which would block the rest of the page from rendering until the LDClient is loaded. documentation says this may sometimes tkae up to 200 ms.
instead, i just wrapped the LDClient code in an if statement. i took a look at the implementation of withLDClient, and it uses React context. so, when the LDClient does load, this should cause the vizier page to rerender and properly start up the LDClient.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6166

GitOrigin-RevId: 044c7890105f73520a68260ea1dd36c1d28eb8e8