PP-2115: Update query broker timeout conditions to support sparse streaming queries.
Summary: Depends on D6173. Streaming queries will often have sparse data and not emit row batches for a while. However, we still want to timeout the query on certain conditions. This diff changes the query broker to timeout if any given result sink takes too long to establish a connection. However, once connections are healthy and established, there is no timeout condition between query results. Continuing the pattern of using agent stream health as a proxy for query health, if an agent result stream closes before it sends its end of stream, we will consider that to be an unexpected closure and treat it as an error.

Test Plan: added units

Reviewers: michelle, zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2115

Differential Revision: https://phab.corp.pixielabs.ai/D6182

GitOrigin-RevId: 7f3c5347ba00f25eac67914d0ed3001dcd447b77