Etcd batch utils: should limit batches by max num txns and byte size
Summary:
some of our customers clusters were hitting a problem where the number of bytes we were flushing in the cache was too much in a single txn.
we are currently batching requests by a max number of txns, however it is possible to have really large requests with less than the max number of txns.
this updates our etcd batcher util to take the max number of bytes into account as well.

Test Plan: unit test

Reviewers: zasgar, #engineering, nserrino

Reviewed By: #engineering, nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D6178

GitOrigin-RevId: 12b4cf9a3ee5ce1b082effe34b930cd5dcc21120