libbpf build
Summary:
Trying to ressurect bpftrace build, but it's not so simple.

Looks like it now has a dependency on libbpf, so adding this into our build.

Note that we're taking the libbpf which is packaged as a BCC submodule. This is to avoid version discrepancies in our code base.

Test Plan: Manual

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6193

GitOrigin-RevId: e3726ffeb091f729c52c0052ca63e08c2abba1c1