Add test for custom protobuf sql driver scanner
Summary:
Was debugging an issue where control plane pod status createdAtMs was returning a negative large number.
This didn't turn out to be the issue, but i wrote a test for the scanner for the control plane pod statuses class to see if it was causing the problem by mangling the data.

Test Plan: the test

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6194

GitOrigin-RevId: 01698045cb6476153465a12038091fa61e5de094