Perform defrags from metadata instead of cronjobs
Summary:
This should give us more control and insight into defrag errors, if any.
for now, we start defragging once the etcd instance hits 500MB. after which, we will defrag for every hour unless it drops below 500MB.
we can calibrate this more once we get a better sense of when we should do defrags.

we should also not perform any cache flushes while defragging, so that all of the agent data stays in memory.

Test Plan: shortened the times so that I don't have to actually wait fo ran hour, then ran on skaffold with etcd stateful set and etcd operator.

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6189

GitOrigin-RevId: 35db15ea116605081d2536d9c7e5af5751cb224d