Fix agent deletion/schema update race condition
Summary:
we have a race condition between agent deletions/schema updates.
currently, it is possible for the following to occur:
- agent heartbeats
- agent updates are put into a queue to be processed
- agent update is processing, sees agent is alive
- agent is deleted on a separate thread
- agent thinks agent is alive, so adds a schema for an agent which should no longer exist

to fix this I did a few things:
- refactor agent deletion so that it only occurs in a single place (when agentHandler is quitting). this required some use of channels and blocking so that if agent A with the same hostname as a newly registering agent B is deleted before we try to register agent B.
- move agent schema/process updates out of the singular queue. now they are processed within the onAgentHeartbeat call.

now, if an agent is deleted, it either must finish processing the current onAgentHeartbeat+applying schema/process updates before actually deleting the agent from the metadatastore. likewise, if a heartbeat comes after the agent is already deleted, the AgentHandler has already stopped, so no updates will be made for this agent.

Test Plan: ran skaffold with a bunch of logs to check things are done in the right order

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6195

GitOrigin-RevId: 346e26e3877fddea2d8a205be2d3601c70ba50eb