Clean up agentHandler Stop logic
Summary:
cleaned up the agentHandler stop logic to use waitgroups rather than the quitDone channel.
also updated the logic to ensure that we cant try writing to the quitCh when it is already closed.
the following is something that could happen:
- Stop() is called because a new agent with X hostname is trying to register and an old agent with X hostname is already registered
- stop() closes the quitCh
- Agent dies before it is registered, for whatever reason
- New agent starts up, Stop() is called again for agent with X hostname
- Stop() tries to write to the quitCh
- panic

Test Plan: ran skaffold, deleted PEMs to confirm that things correctly get deleted/registered

Reviewers: zasgar, #engineering, nserrino

Reviewed By: #engineering, nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D6199

GitOrigin-RevId: 05abab4144f55c521900bc5df83f0ae064181542