Allow agent manager to support concurrent agent update stream consumers
Summary:
We have a problem between the query broker and the metadata service across the API GetAgentUpdates. This API was originally designed for a single consumer (the singleton query broker).

We are running into an issue where the stream between them times out from the query broker's perspective. As a result, the query broker decides to reconnect. However, despite the GRPC error on the query broker side, from the metadata service's perspective, for some reason the first stream stays alive and the second stream gets connected too. GetAgentUpdates was only designed to support a single consumer, so both streams update the state, leading to inconsistent results from the non-zombie stream that has reconnected on the query broker.

@michelle is looking into the zombie streams issue, but for now this change allows the metadata services GetAgentUpdates to support multiple consumers. This is a step that is necessary for us anyway once we want to support multiple mds and qbs, and should put us a bit closer to the eventual design with versioned updates.

Test Plan: edited existing

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

Subscribers: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D6202

GitOrigin-RevId: 0830feef6adddbe7e105ba34446e3fd04ca3040c