Return VarType::kUnspecified to indicate an unsupported type
Summary:
Previously the function returns an error status.

This avoids confusing error when deploying a probe that does not
actually probe arguments with unsupported types.

GetFunctionArgInfo() is called to return the ArgInfo of all of the
arguments of a function, regardless of whether the argument is
probed or not.

This fix tries to minimize the scope of changes.

A possible alternative is to only resolve ArgInfo for each arg
expression, which appears more intrusive.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6203

GitOrigin-RevId: 6f18e52e97d0be957b1c038e0229ca4557f9b1f6