OpenSSL tracing fix
Summary: OpenSSL probes were not deploying properly on k8s because stirling/PEM is inside a container and paths need to be adjusted appropriately.

Test Plan: Manual. We already have a test with nginx inside a container. Will leave the exercise of stirling being inside a container as a separate task.

Reviewers: zasgar, yzhao, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6218

GitOrigin-RevId: 356d537dfa4b7e2685cb3b56882a76b04fffee96