[PP-2180] GRPC Bridge Pushdown Optimization in GRPCSplitter
Summary:
GRPCBridges are an expensive operation in Vizier.  Complicated queries ended up having many GRPCBridges that were sending duplicate/overlapping data. That meant our network bandwidth was much higher than it needed to be.

To fix this, we can do a form of GRPCBridge push-down when we find that data sent over is highly duplicated. The process applies a simple heuristic:
For each source in the logical plan, we find all the GRPCBridges that depend on that source. If any of those GRPCBridges do not use a partial operator, then we merge all of the GRPCBridges together into one and push it down to the first "branch" of the query.

Of course this makes a number of assumptions that have edge conditions invalidating the heuristic
1. There's a limit of # of partial blcoking operations together might actually produce more data/network load than a single GRPCBridge with all of the source data.
2. We don't consider filters at all when making this push down. It could be that every branch has a different filter that ends up reducing the sum network below the value of a single stream.

And potentially others. I've avoided addressing these because these require estimates of the physical state of the data which we do not have the data for at the moment.

Test Plan: Wrote new tests and tested that it still works on GKE.

Reviewers: nserrino, jamesbartlett, zasgar, #engineering

Reviewed By: nserrino, #engineering

JIRA Issues: PP-2180

Differential Revision: https://phab.corp.pixielabs.ai/D6133

GitOrigin-RevId: 7c23212e0cbb71a2e3cbf15333c3f0f500144c87