PP-2115: Change the way timeouts work in Carnot's exec graph.
Summary:
With the addition of streaming to Vizier, we are going to have queries that only send data very occasionally. This breaks the concept of timeouts as they are handled in Carnot today. Carnot times out if it hasn't received data from a remote source after a certain period of time. In the streaming model, not receiving data from a remote source for a while is a perfectly legitimate case for a query on sparse data. As a result, we want to remove the data timeout and add 2 other checks. 1. If the connection between Carnot instances becomes closed while the source still expects data from the sink, that triggers an execution error during the query. 2. If a remote sink takes too long to establish the connection to its Carnot destination, then that is the new timeout condition evaluated in the exec graph.

Next up is adjustments to the query broker to do similar handling and remove data-based timeouts.

Depends on D6169.

Test Plan: added more tests.

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-2115

Differential Revision: https://phab.corp.pixielabs.ai/D6173

GitOrigin-RevId: 9ea21300ab3b4103fcb6b8b9111bc6c00a488f83