Update etcd to retain less snapshots/wal
Summary:
suggestions on the etcd github issues suggest to further reduce the snapshot count, which will increase snapshot frequency and keep less raft logs in memory.
we can also reduce the number of wals/snapshots we keep around.
for example, if the etcd instance is taking 1.2 GB of actual space, if we keep 2 snapshots around that will be 3.6 GB of space.
there should also be a separate investigation to reduce the amount of space we actually use for our data on etcd.

Test Plan: created rc and verified things still run

Reviewers: zasgar, #engineering, nserrino

Reviewed By: #engineering, nserrino

Differential Revision: https://phab.corp.pixielabs.ai/D6235

GitOrigin-RevId: 7469444037573588d06e7630b0811fc1a9f6d7db