Metadata GetTracepointInfo should show better errors
Summary:
before, we were only showing the first agent's error, which may not show the relevant error.
instead, we should filter out all errors that are FAILED_PRECONDITION, since those are likely to occur. we only show the FAILED_PRECONDITION error if there are no PEMs which match the given UPID.
we also send an array of statuses rather than a single status, to prepare for the future where the tracepoint may be deployed on multiple PEMS.

Test Plan: ran on skaffold, confirm that i see more interesting errors

Reviewers: philkuz, zasgar, #engineering

Reviewed By: philkuz, #engineering

Subscribers: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D6253

GitOrigin-RevId: 77f8dafd41859cb042e0d45f2affdbf8cd29bdc9