PP-2115 Ensure that cancellation of queries by query broker propagates to PEMs
Summary:
When the query broker cancels a query, it terminates the streams between Kelvin and itself, thus causing Kelvin to know that the query is cancelled. However, there is a layer of indirection between the PEMs and Kelvin for this process. PEMs talk to the Kelvin's GRPC router, which is not query specific. If PEMs send a result stream to a Kelvin for a query that doesn't exist, this is interpreted by Kelvin as data to store in case the query will show up later. This causes a problem for the PEMs, who never find out they need to cancel the query, because the stream is never shut down like it is for the Kelvin<->QB case.

My workaround here is to store the server contexts for all of the inbound streams received by the GRPC router. Then, when a query is deleted as the result of a cancellation, those server contexts will be cancelled, thus aborting the TransferResultChunk streams associated with them.

We still have a problem in the current implementation where the agents will not realize that their stream is closed until they attempt a write to that stream in the GRPC Sink. Ideally we would determine that the stream is closed earlier than that point, and abort the query. This will be especially problematic if someone eventually writes a windowed query, where the window is 1h. In that scenario, since data is written only once per hour, it would potentially take up to an hour for an agent to be aware that the query is actually cancelled (when it finally attempts the write). We need to either 1. figure out how to determine that a stream is closed from the client side (this is possible from the grpc::ServerContext with its IsCancelled() method, but the grpc ClientContext does not appear to have such a method) without relying on a write, or 2. have a separate thread/task which sends heartbeats to all downstream destinations, and when those fail, cancel the corresponding query. #1 would be more efficient of a solution but it is unclear how to do that in C++ (despite it being easy to do in Golang).

However, the above issue won't be a problem until we have windowed streaming queries, which we don't at this time.

Test Plan: ran a query that previously left a zombie in the pem, no longer leaves a zombie in the pem.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2115

Differential Revision: https://phab.corp.pixielabs.ai/D6257

GitOrigin-RevId: fde807b42d590d757acd8a7eae4c6462b840f3d1