Adding support for bpftrace strings
Summary: Added support for bpftrace strings. Will need to be updated when more diffs land for the proto changes.

Test Plan: added a test to mak sure we parse the strings as expected. Discovered that python normally parses \n rather  than preserving it. so we have to escape them or change the parser to support it.

Reviewers: oazizi, yzhao, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6265

GitOrigin-RevId: 336ecd1757a705cd49741c9f2e262e76f089f858