PP-2115 Support streaming queries in CLI for px run.
Summary: If someone does px run with a streaming query, we should support it if their output is json. Streaming won't work with a normal table output because you cannot print a complete streaming table all in one go. When all of our live views become streaming, I think for the table output mode for px run, we can just set an end_time in the query with a warning to the user that it won't stream the data in the px run -o table mode. However, since JSON data prints out the table name for every record, it can be used to stream results back today. Needed to add ability to CTRL+C a query so that infinite streaming queries can be stopped by the user.

Test Plan: ran streaming queries and cancelled them

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2115

Differential Revision: https://phab.corp.pixielabs.ai/D6275

GitOrigin-RevId: 65b70918f4dfa419feb02865210196cb2d129978