PP-2115: Flush data from unions on a timeout
Summary: Previously, we buffered data in unions until it reached 1024 output rows or end of stream. Now with the introduction of streaming, that will lead to a long latency for data to show up in results if the query only produces a few rows at a time. As a result, I added a timeout to union that flushes ready data every second so that we still buffer within a certain timeframe but flush if it has been a while.

Test Plan: added a unit test for the timeout and no timeout cases.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2115

Differential Revision: https://phab.corp.pixielabs.ai/D6259

GitOrigin-RevId: 2d42746a193b568ff417661d9beea819f8420037