Update graph styling to reflect edge/node weight/color updates
Summary:
we want to allow for more customization of the graph widget visualizations, such as color/weight of edges and nodes.
the previous diff updated the vis spec proto. this diff actually implements the changes in the graph itself

Test Plan:
ran webpack and modified the network flow graph vis spec:
{F95208}

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6285

GitOrigin-RevId: bdc84ed882dea4b0cc28b3c3e101e35312b3ca16