Add implementation of k-means.
Summary: Implements k-means in Eigen, using kmeans++ as the initialization algorithm. In the future, I probably want to change this to be a mini-batch style kmeans, but since we're using the coreset tree to reduce the data anyways it probably doesn't matter too much.

Test Plan: Added test for basic case, tried out in end-2-end demo.

Reviewers: #engineering, zasgar

Reviewed By: #engineering, zasgar

Subscribers: zasgar, philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D6293

GitOrigin-RevId: 6ef579c3122c94b9227ac9c72725dba4b9d61e1a