Add ModelPool to allow UDFs/UDAs to borrow an initialized model.
Summary:
This adds the ModelPool which allows UDFs to borrow a pointer to a pre-initialized model. Right now, the pool will only create one model, and allow calls must share that copy, but this can be changed in the future if we need more parallel queries.

This diff also implements inference for the transformer style model.

@zasgar We talked about this awhile back but I think I strayed a bit from what we were talking about, so would be good to hear your thoughts on this one.

Test Plan: Added some tests and used in end2end

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Subscribers: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D6294

GitOrigin-RevId: f34c22892644f413025bf4e7973b5a8e25f3d07a