Add pem_only option to LimitIR.
Summary:
This diff adds an option to df.head, to specify to keep the limit on the PEMs only. The syntax is: `df.head(n, pem_only=1)`
This allows me to run the models on the pems but also limit the amount of data each pem runs on.

Test Plan: Added tests

Reviewers: philkuz, nserrino, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6297

GitOrigin-RevId: 3612e7bc75a4d289ae50b9521fe300363443ecf1