PP-2115: GRPCSinks should periodically send an empty row batch to their downstream consumers to ensure a query cancellation is detected
Summary: We currently have an issue with streaming queries, where pems don't realize the query has been cancelled until they send data to Kelvin, which then responds that the query is cancelled. This diff fixes that by supporting checking the grpc connection to the downstream recipient of the grpc sink by sending an empty row batch. If the connection has been closed, this will fail, and the agent will know that the query should be cancelled. Ideally we would do this by checking the connection status (rather than sending an empty batch) but this doesn't appear to be possible in grpc::ClientContext. This logic will likely get subsumed by either 1. periodic exec stats updates 2. a change to the way streaming queries work where all streams are inherently windowed, and send an empty batch per window even if there is no data there. But for now, this should fix the problem where a pem that never sends any data never discovers the query is cancelled, thus leading to a zombie.

Test Plan: unit

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2115

Differential Revision: https://phab.corp.pixielabs.ai/D6307

GitOrigin-RevId: 6b777e8b7b5c0b5ad86bcb339349a9e991c88445