Fix data details in UI
Summary:
the diff which updated the UI to work for streaming added some more dependencies to the React hooks.
this resulted in some additional hooks to be called, which broke the data details row selection.
we need to do some react callbacks so that the hooks don't get triggered when the table is just rerendering.

Test Plan: ran webpack, tried opening it. also made sure sorting with/without streaming still works.

Reviewers: nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6315

GitOrigin-RevId: bc1e1039d838865240d680ab5fca164a409f42d2