Correct scrolling behavior in Admin Overview page.
Summary:
Shuffles the structure of Admin Overview such that there is one scrollbar, and the tab bar sticks to the top of the containing viewport. Also switches two components from `withStyles` to `useStyles` so that TypeScript correctly infers the classes contained within.

{F95594}

Test Plan: Visit the Admin Overview page, on the Deployment Keys tab. There should be a long table present. Scrolling to the bottom should reach the end of the table, and there should only be one scrollbar. Real mobile still needs to be tested.

Reviewers: #engineering, nserrino, michelle

Reviewed By: #engineering, nserrino

JIRA Issues: PC-552

Differential Revision: https://phab.corp.pixielabs.ai/D6324

GitOrigin-RevId: 602bfe4b9a9aad5475f2d52460afb13953798ca7