PC-525: Don't show raw UPID protobuf in row details inspector
Summary:
Previously we would show a complex protobuf representation of uint128 data in the row details inspector. This was not what the user wants, which is just the uint128 in string form. Now, when a user clicks on a row, it will be easier for them to interpret and use upids. I convert the uint128 protobuf when we load the result data, since the protobuf form is not needed or wanted in the UI.

Also, I noticed some issues with sorting upids and container state/pod state semantic types, so I fixed those as well in this change.

Test Plan: edited unit tests.

Reviewers: nick, michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-525

Differential Revision: https://phab.corp.pixielabs.ai/D6325

GitOrigin-RevId: e84c8f76570dead18e9f91c6b566fd30022d2fc5