Fix "script is already executing" in UI
Summary:
with my streaming UI changes, we cancel the previous query before starting a new one...
this actually led to a small bug, because I'd left in code that checked whether the previous query was still loading. if it was, we would show a popup notifying the user the previous query was still executing, and wouldn't execute the new query.
but since the previous query had been cancelled, the query will never actually complete and they would get stuck in this state.
we should just not show this popup and just let the user execute the new query.

Test Plan: ran in webpack. run px/cluster in customer, then switched to another script while that was running. confirmed that the new results load, and the previous query was cancelled.

Reviewers: nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6337

GitOrigin-RevId: 82c411591faf8448f441ede41adc4df87d0b867f