Table store: support multiple sources to a single table.
Summary:
When a new dynamic trace is deployed, a brand new data source is created, but it may want to push to an existing table.

Right now, the data just goes into a black hole.

This is also affects the scenario where a dynamic data source is destroyed and then redeployed later. If it specifies the old table, appends currently don't work.

This diff adds support for a new data source to push to an old table.

Test Plan: Test added to table store.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2222

Differential Revision: https://phab.corp.pixielabs.ai/D6339

GitOrigin-RevId: 03e0b56ed8e4b9e2402cbcbb7bd3190e3addc977