PC-561 Allow arbitrary strings in breadcrumbs and fix focusing
Summary:
recent refactors made it so that we could not submit arbitrary strings in the breadcrumbs.
this diff updates it so that some breadcrumbs may require the user to select from the suggestions, while others do not.

also fixed the focus, which required passing a ref through to the component that actually contains the HTML input element.

Test Plan: ran with webpack, enter arbitrary strings and press enter.

Reviewers: nick, nserrino, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6349

GitOrigin-RevId: 6d6be7199470554557ba001488954e100ec9e565