[PP-2223] Fix Dependent operators_to_remove causes a segfault
Summary: Operators to remove sometimes remove one filter when removing the other operators. Before this would cause a segfault. This diff adds a test and a fix. Will follow up with unit tests of the new coordinator pieces.

Test Plan: Tested and seems to work with the specific failing case.

Reviewers: nserrino, zasgar, jamesbartlett, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2223

Differential Revision: https://phab.corp.pixielabs.ai/D6354

GitOrigin-RevId: 5d42db97050a7c373ffa007de7b01ba9df00a4c7