Fix bug where updates to existing schemas don't get picked up in metadata store, this time without error log.
Summary:
This was preventing ST_BYTES/ST_DURATION_NS from showing up as semantic types on http_table if etcd hadn't been clobbered first.
This is the same diff as D6358 but removed the error log.

Test Plan:
added a unit test that failed before the changes. also ran a script and saw the semantic type was correct.
Ran this and D6358 on @oazizi's cluster, and confirmed that D6358 didn't work but this version does.

Reviewers: #engineering, philkuz, michelle, zasgar

Reviewed By: #engineering, zasgar

Subscribers: oazizi

Differential Revision: https://phab.corp.pixielabs.ai/D6391

GitOrigin-RevId: 75f83e11ce784f2f27044e8a8db7dbcca0f7b88f