Add semantic type formatting for duration and bytes
Summary:
This adds formatting to the table. Next is Vega and Maps.

We need to think about refactoring this code to make this cleaner/easier after we move to Rx. Rough idea:

We always keep around a data structure which refers to the type, value. We can then layer on formatting to capture cases like
color, value formatting, etc.

Test Plan: visual check, didn't write tests because I expect this code to go away soon when we move to Rx.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6372

GitOrigin-RevId: e96b3675a9947f0aa55efba2d0ba0e391d545b0d