Fix text on login pages
Summary:
because we added a typography to our mui-theme, the default typography classes which we didn't specify (subtitle1, subtitle2, etc) no longer work.
I redefined them in the mui-themes with the values that they had previously so that the login page looks the same, using the UI running in staging as reference.

Test Plan: ran in webpack

Reviewers: zasgar, nserrino, nick, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D6369

GitOrigin-RevId: 5ad13de1da2378941c391af85ff447c516bb5cba