Show the real input with the real caret in breadcrumbs autocomplete. Style to match.
Summary:
This allows the user to treat the input field as an input field, with a working selection caret. Appearance unchanged; suggestions still appear as if continuing the user's input. This is achieved with some wonky CSS, but removes the even wonkier custom caret.

A side effect of this is that colors within the typed segment are no longer possible (this was not being done with this component - `new-autocomplete` is being used separately for that in the Command input). That functionality could be easily restored via another overlay, but only if the coloring does not affect font size/weight/spacing.

Demo (screenshot):

{F96076}

Test Plan: Use the breadcrumb autocomplete and notice that shift+arrows, Ctrl/Cmd+A, and mouse usage now behave as they do in a normal input, while the appearance is unaffected and suggestions still appear.

Reviewers: nserrino, michelle, #engineering, zasgar

Reviewed By: #engineering, zasgar

JIRA Issues: PC-592

Differential Revision: https://phab.corp.pixielabs.ai/D6386

GitOrigin-RevId: 8482b04d29ee39e18915ff31a6505a22229dab1a